requirements.txt
attrs==25.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
flasgger==0.9.7.1
flask==3.1.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
packaging==25.0
pillow==11.2.1
PyMySQL==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
sqlalchemy==2.0.40
typing-extensions==4.13.0
werkzeug==3.1.3
zipp==3.21.0
